Lookup Page
Started to design the lookup page. This will be the default page that
the user lands on. There is still plenty of work to be done, making the
form, and creating the JavaScript to update the output area. This whole
page can be swapped out by making the display of `lookup` equal to none.